+++ scripts/macros/2024/feats/elementalAdept.js
@@ -1,0 +1,26 @@
+import {elementalAdeptA as elementalAdeptALegacy, elementalAdeptC as elementalAdeptCLegacy, elementalAdeptF as elementalAdeptFLegacy, elementalAdeptL as elementalAdeptLLegacy, elementalAdeptT as elementalAdeptTLegacy} from '../../../legacyMacros.js';
+export let elementalAdeptA = {
+    ...elementalAdeptALegacy,
+    version: '1.2.36',
+    rules: 'modern'
+};
+export let elementalAdeptC = {
+    ...elementalAdeptCLegacy,
+    version: elementalAdeptA.version,
+    rules: elementalAdeptA.rules
+};
+export let elementalAdeptF = {
+    ...elementalAdeptFLegacy,
+    version: elementalAdeptA.version,
+    rules: elementalAdeptA.rules
+};
+export let elementalAdeptL = {
+    ...elementalAdeptLLegacy,
+    version: elementalAdeptA.version,
+    rules: elementalAdeptA.rules
+};
+export let elementalAdeptT = {
+    ...elementalAdeptTLegacy,
+    version: elementalAdeptA.version,
+    rules: elementalAdeptA.rules
+};