+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.36 Release:
 ## New Modern Content:
 - Crusher
+- Elemental Adept
 
 # 1.2.35 Release:
 ## Bug Fixes: